+++ src/main/java/io/github/talelin/latticy/dto/admin/DispatchPermissionDTO.java
@@ -7,15 +7,16 @@ import javax.validation.constraints.Positive;
 
 /**
  * @author pedro@TaleLin
+ * @author Juzi@TaleLin
  */
 @Data
 public class DispatchPermissionDTO {
 
     @Positive(message = "{group.id.positive}")
     @NotNull(message = "{group.id.not-null}")
-    private Long groupId;
+    private Integer groupId;
 
     @Positive(message = "{permission.id.positive}")
     @NotNull(message = "{permission.id.not-null}")
-    private Long permissionId;
+    private Integer permissionId;
 }
