+++ src/main/java/io/github/talelin/latticy/dto/admin/DispatchPermissionsDTO.java
@@ -9,14 +9,15 @@ import java.util.List;
 
 /**
  * @author pedro@TaleLin
+ * @author Juzi@TaleLin
  */
 @Data
 public class DispatchPermissionsDTO {
 
     @Positive(message = "{group.id.positive}")
     @NotNull(message = "{group.id.not-null}")
-    private Long groupId;
+    private Integer groupId;
 
     @LongList(message = "{permission.ids.long-list}")
-    private List<Long> permissionIds;
+    private List<Integer> permissionIds;
 }
