+++ src/main/java/io/github/talelin/latticy/service/GroupService.java
@@ -1,15 +1,16 @@
 package io.github.talelin.latticy.service;
 
 import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.extension.service.IService;
 import io.github.talelin.latticy.bo.GroupPermissionBO;
 import io.github.talelin.latticy.common.enumeration.GroupLevelEnum;
 import io.github.talelin.latticy.model.GroupDO;
-import com.baomidou.mybatisplus.extension.service.IService;
 
 import java.util.List;
 
 /**
  * @author pedro@TaleLin
+ * @author Juzi@TaleLin
  */
 public interface GroupService extends IService<GroupDO> {
 
@@ -19,7 +20,7 @@ public interface GroupService extends IService<GroupDO> {
      * @param userId 用户id
      * @return 所有分组
      */
-    List<GroupDO> getUserGroupsByUserId(Long userId);
+    List<GroupDO> getUserGroupsByUserId(Integer userId);
 
     /**
      * 获得用户的所有分组id
@@ -27,7 +28,7 @@ public interface GroupService extends IService<GroupDO> {
      * @param userId 用户id
      * @return 所有分组id
      */
-    List<Long> getUserGroupIdsByUserId(Long userId);
+    List<Integer> getUserGroupIdsByUserId(Integer userId);
 
     /**
      * 分页获取分组数据
@@ -36,7 +37,7 @@ public interface GroupService extends IService<GroupDO> {
      * @param page  那一页
      * @return 分组页
      */
-    IPage<GroupDO> getGroupPage(long page, long count);
+    IPage<GroupDO> getGroupPage(int page, int count);
 
     /**
      * 通过id检查分组是否存在
@@ -44,7 +45,7 @@ public interface GroupService extends IService<GroupDO> {
      * @param id 分组id
      * @return 是否存在
      */
-    boolean checkGroupExistById(Long id);
+    boolean checkGroupExistById(Integer id);
 
     /**
      * 获得分组及其权限
@@ -52,7 +53,7 @@ public interface GroupService extends IService<GroupDO> {
      * @param id 分组id
      * @return 分组及权限
      */
-    GroupPermissionBO getGroupAndPermissions(Long id);
+    GroupPermissionBO getGroupAndPermissions(Integer id);
 
     /**
      * 通过名称检查分组是否存在
@@ -68,7 +69,7 @@ public interface GroupService extends IService<GroupDO> {
      * @param userId 用户id
      * @return true表示在
      */
-    boolean checkIsRootByUserId(Long userId);
+    boolean checkIsRootByUserId(Integer userId);
 
     /**
      * 删除用户与分组直接的关联
@@ -76,7 +77,7 @@ public interface GroupService extends IService<GroupDO> {
      * @param userId    用户id
      * @param deleteIds 分组id
      */
-    boolean deleteUserGroupRelations(Long userId, List<Long> deleteIds);
+    boolean deleteUserGroupRelations(Integer userId, List<Integer> deleteIds);
 
     /**
      * 添加用户与分组直接的关联
@@ -84,7 +85,7 @@ public interface GroupService extends IService<GroupDO> {
      * @param userId 用户id
      * @param addIds 分组id
      */
-    boolean addUserGroupRelations(Long userId, List<Long> addIds);
+    boolean addUserGroupRelations(Integer userId, List<Integer> addIds);
 
     /**
      * 获得分组下所有用户的id
@@ -92,8 +93,7 @@ public interface GroupService extends IService<GroupDO> {
      * @param id 分组id
      * @return 用户id
      */
-    List<Long> getGroupUserIds(Long id);
-
+    List<Integer> getGroupUserIds(Integer id);
 
     /**
      * 通过分组级别获取超级管理员分组或游客分组
@@ -109,5 +109,5 @@ public interface GroupService extends IService<GroupDO> {
      * @param level GroupLevelEnum 枚举类
      * @return 用户组id
      */
-    Long getParticularGroupIdByLevel(GroupLevelEnum level);
+    Integer getParticularGroupIdByLevel(GroupLevelEnum level);
 }
