+++ src/main/java/io/github/talelin/latticy/service/impl/GroupServiceImpl.java
@@ -2,21 +2,19 @@ package io.github.talelin.latticy.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
-import io.github.talelin.autoconfigure.exception.NotFoundException;
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import io.github.talelin.autoconfigure.exception.ForbiddenException;
 import io.github.talelin.latticy.bo.GroupPermissionBO;
 import io.github.talelin.latticy.common.enumeration.GroupLevelEnum;
 import io.github.talelin.latticy.common.mybatis.Page;
+import io.github.talelin.latticy.mapper.GroupMapper;
 import io.github.talelin.latticy.mapper.UserGroupMapper;
 import io.github.talelin.latticy.model.GroupDO;
-import io.github.talelin.latticy.mapper.GroupMapper;
 import io.github.talelin.latticy.model.PermissionDO;
 import io.github.talelin.latticy.model.UserGroupDO;
 import io.github.talelin.latticy.service.GroupService;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import io.github.talelin.latticy.service.PermissionService;
-import io.github.talelin.autoconfigure.exception.ForbiddenException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -37,28 +35,28 @@ public class GroupServiceImpl extends ServiceImpl<GroupMapper, GroupDO> implemen
     private UserGroupMapper userGroupMapper;
 
     @Override
-    public List<GroupDO> getUserGroupsByUserId(Long userId) {
+    public List<GroupDO> getUserGroupsByUserId(Integer userId) {
         return this.baseMapper.selectUserGroups(userId);
     }
 
     @Override
-    public List<Long> getUserGroupIdsByUserId(Long userId) {
+    public List<Integer> getUserGroupIdsByUserId(Integer userId) {
         return this.baseMapper.selectUserGroupIds(userId);
     }
 
     @Override
-    public IPage<GroupDO> getGroupPage(long page, long count) {
+    public IPage<GroupDO> getGroupPage(int page, int count) {
         Page pager = new Page(page, count);
         return this.baseMapper.selectPage(pager, null);
     }
 
     @Override
-    public boolean checkGroupExistById(Long id) {
+    public boolean checkGroupExistById(Integer id) {
         return this.baseMapper.selectCountById(id) > 0;
     }
 
     @Override
-    public GroupPermissionBO getGroupAndPermissions(Long id) {
+    public GroupPermissionBO getGroupAndPermissions(Integer id) {
         GroupDO group = this.baseMapper.selectById(id);
         List<PermissionDO> permissions = permissionService.getPermissionByGroupId(id);
         return new GroupPermissionBO(group, permissions);
@@ -72,9 +70,9 @@ public class GroupServiceImpl extends ServiceImpl<GroupMapper, GroupDO> implemen
     }
 
     @Override
-    public boolean checkIsRootByUserId(Long userId) {
+    public boolean checkIsRootByUserId(Integer userId) {
         QueryWrapper<UserGroupDO> wrapper = new QueryWrapper<>();
-        Long rootGroupId = this.getParticularGroupIdByLevel(GroupLevelEnum.ROOT);
+        Integer rootGroupId = this.getParticularGroupIdByLevel(GroupLevelEnum.ROOT);
         wrapper.lambda().eq(UserGroupDO::getUserId, userId)
                 .eq(UserGroupDO::getGroupId, rootGroupId);
         UserGroupDO relation = userGroupMapper.selectOne(wrapper);
@@ -82,7 +80,7 @@ public class GroupServiceImpl extends ServiceImpl<GroupMapper, GroupDO> implemen
     }
 
     @Override
-    public boolean deleteUserGroupRelations(Long userId, List<Long> deleteIds) {
+    public boolean deleteUserGroupRelations(Integer userId, List<Integer> deleteIds) {
         if (deleteIds == null || deleteIds.isEmpty()) {
             return true;
         }
@@ -97,7 +95,7 @@ public class GroupServiceImpl extends ServiceImpl<GroupMapper, GroupDO> implemen
     }
 
     @Override
-    public boolean addUserGroupRelations(Long userId, List<Long> addIds) {
+    public boolean addUserGroupRelations(Integer userId, List<Integer> addIds) {
         if (addIds == null || addIds.isEmpty()) {
             return true;
         }
@@ -110,7 +108,7 @@ public class GroupServiceImpl extends ServiceImpl<GroupMapper, GroupDO> implemen
     }
 
     @Override
-    public List<Long> getGroupUserIds(Long id) {
+    public List<Integer> getGroupUserIds(Integer id) {
         QueryWrapper<UserGroupDO> wrapper = new QueryWrapper<>();
         wrapper.lambda().eq(UserGroupDO::getGroupId, id);
         List<UserGroupDO> list = userGroupMapper.selectList(wrapper);
@@ -130,12 +128,12 @@ public class GroupServiceImpl extends ServiceImpl<GroupMapper, GroupDO> implemen
     }
 
     @Override
-    public Long getParticularGroupIdByLevel(GroupLevelEnum level) {
+    public Integer getParticularGroupIdByLevel(GroupLevelEnum level) {
         GroupDO group = this.getParticularGroupByLevel(level);
-        return group == null ? 0L : group.getId();
+        return group == null ? 0 : group.getId();
     }
 
-    private boolean checkGroupExistByIds(List<Long> ids) {
+    private boolean checkGroupExistByIds(List<Integer> ids) {
         return ids.stream().allMatch(this::checkGroupExistById);
     }
 }
